backend/requirements.txt
# Django 核心與 API
Django==4.1.7
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework-simplejwt==5.3.0

# 資料庫連線 (MySQL)
mysqlclient==2.2.4
mysql-connector-python==8.3.0

# 環境變數管理
python-dotenv==1.2.1

# 其他輔助
requests==2.31.0
asgiref==3.6.0
sqlparse==0.4.3
pytz==2022.7.1